Refactor cookie banner
