+++ app/cookie_banner.py
@@ -1,0 +1,59 @@
+import dash
+from dash import html, dcc
+
+store = dcc.Store(id="cookie-consent-store", storage_type="local")
+
+layout = html.Div(
+    id="cookie-banner",
+    children=[
+        html.Div(
+            [
+                html.P(
+                    "This website uses cookies to improve your experience.",
+                    className="mb-0 me-3",
+                ),
+                html.Button(
+                    "Accept",
+                    id="accept-cookies",
+                    className="btn btn-success me-2",
+                ),
+                html.Button("Reject", id="reject-cookies", className="btn btn-danger"),
+            ],
+            className="alert alert-dark d-flex justify-content-center align-items-center gap-2 px-3 py-2",
+            style={
+                "position": "fixed",
+                "bottom": "0",
+                "width": "100%",
+                "zIndex": "1050",
+                "borderRadius": "0",
+            },
+        )
+    ],
+    hidden=True,
+)
+
+
+def register_callbacks(app):
+    @app.callback(
+        [
+            dash.Output("cookie-banner", "hidden"),
+            dash.Output("cookie-consent-store", "data"),
+        ],
+        [
+            dash.Input("accept-cookies", "n_clicks"),
+            dash.Input("reject-cookies", "n_clicks"),
+        ],
+        [dash.State("cookie-consent-store", "data")],
+        prevent_initial_call=False,
+    )
+    def handle_cookie_consent(accept_clicks, reject_clicks, store_data):
+        ctx = dash.callback_context
+        if store_data:
+            return True, dash.no_update  # Hide banner if choice was already made.
+
+        if ctx.triggered:
+            button_id = ctx.triggered[0]["prop_id"].split(".")[0]
+            if button_id in ["accept-cookies", "reject-cookies"]:
+                return True, {"cookie_consent": button_id}
+
+        return False, dash.no_update  # Show banner if no choice stored.
